refactor watcher
